+++ test/frontend/require.js
@@ -15,9 +15,9 @@ window.require = function(name) {
     return Q;
   }
 
-  // chai-datetime is self-registering in the browser
+  // chai plugins are self-registering in the browser
   // we're returning a noop function so that the call to chai.use() does nothing
-  if (name === 'chai-datetime') {
+  if (name === 'chai-datetime' || name === 'chai-shallow-deep-equal') {
     return function() {};
   }
 };